メンター以外がコメントしたときに担当者になる問題を修正
Fix: #4146
